+++ apps/pdv-backend/src/modules/fiscal/fiscal.module.ts
@@ -8,6 +8,7 @@ import { FISCAL_GATEWAY } from "./application/ports/fiscal-gateway.port.js";
 import { GetFiscalStatusUseCase } from "./application/use-cases/get-fiscal-status.use-case.js";
 import { IssueFiscalReceiptManuallyUseCase } from "./application/use-cases/issue-fiscal-receipt-manually.use-case.js";
 import { RetryFiscalDocumentUseCase } from "./application/use-cases/retry-fiscal-document.use-case.js";
+import { ReissueFiscalDocumentUseCase } from "./application/use-cases/reissue-fiscal-document.use-case.js";
 
 @Module({
   imports: [ProvisioningModule],
@@ -16,6 +17,7 @@ import { RetryFiscalDocumentUseCase } from "./application/use-cases/retry-fiscal
     GetFiscalStatusUseCase,
     IssueFiscalReceiptManuallyUseCase,
     RetryFiscalDocumentUseCase,
+    ReissueFiscalDocumentUseCase,
     { provide: FISCAL_DOCUMENT_REPOSITORY, useClass: PrismaFiscalDocumentRepository },
     { provide: FISCAL_GATEWAY, useClass: HttpFiscalGateway },
   ],